backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
click==8.2.1
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.115.0
h11==0.16.0
httptools==0.6.4
idna==3.10
pydantic==2.9.2
pydantic-core==2.23.4
pymongo==4.8.0
python-dotenv==1.1.1
python-multipart==0.0.9
PyYAML==6.0.2
sniffio==1.3.1
starlette==0.38.6
typing_extensions==4.15.0
uvicorn==0.30.6
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
